Moved prepare-request to sepeareate file for more easy testing. Refactoring
