+++ package.json
@@ -58,8 +58,8 @@
     "eslint-plugin-react": "^6.4.1",
     "isparta": "^4.0.0",
     "mocha": "^1.21.5",
-    "proxyquire": "^1.7.11",
-    "sinon": "^1.17.6",
-    "sinon-chai": "^2.8.0"
+    "proxyquire": "^1.8.0",
+    "sinon": "^1.17.7",
+    "sinon-chai": "^2.12.0"
   }
 }
